Contact page and navbar
Created a contact page with a simple contact form for users to easily submit questions to the given store. page also has general contact infomation.

Added a simple navbar at the top of the page, for users to sign up/log in and/or check additional information about their own account and history.